ReviewStatus enum now loaded directly into db file and enums used are delegated to a seperate file.
