+++ cedars/app/ops.py
@@ -28,8 +28,8 @@ from . import auth
 from .database import minio
 from .api import load_pines_url, kill_pines_api
 from .api import get_token_status
-from .adjudication_handler import AdjudicationHandler, PatientStatus
-from .adjudication_handler import ReviewStatus
+from .adjudication_handler import AdjudicationHandler
+from .cedars_enums import PatientStatus
 
 
 bp = Blueprint("ops", __name__, url_prefix="/ops")
@@ -563,13 +563,10 @@ def save_adjudications():
         annotations_after_event = []
         if skip_after_event:
             annotations_after_event = db.get_annotations_post_event(patient_id,
-                                                                    new_date,
-                                                                    ReviewStatus.UNREVIEWED)
-            db.mark_annotations_post_event(patient_id, new_date, ReviewStatus.UNREVIEWED,
-                                                                 ReviewStatus.SKIPPED)
+                                                                    new_date)
+            db.mark_annotations_post_event(patient_id, new_date)
 
-        db.mark_annotation_reviewed(current_annotation_id, current_user.username,
-                                    ReviewStatus.REVIEWED, ReviewStatus.UNREVIEWED)
+        db.mark_annotation_reviewed(current_annotation_id, current_user.username)
         db.update_event_date(patient_id, new_date, current_annotation_id)
 
         adjudication_handler.mark_event_date(new_date, current_annotation_id,
@@ -577,14 +574,11 @@ def save_adjudications():
 
     elif action == 'del_date':
         db.delete_event_date(patient_id)
-        db.revert_skipped_annotations(patient_id, ReviewStatus.SKIPPED,
-                                                ReviewStatus.REVIEWED)
-        db.mark_annotation_reviewed(current_annotation_id, current_user.username,
-                                    ReviewStatus.REVIEWED, ReviewStatus.UNREVIEWED)
+        db.revert_skipped_annotations(patient_id)
+        db.mark_annotation_reviewed(current_annotation_id, current_user.username)
         adjudication_handler.delete_event_date()
     elif action == 'adjudicate':
-        db.mark_annotation_reviewed(current_annotation_id, current_user.username,
-                                    ReviewStatus.REVIEWED, ReviewStatus.UNREVIEWED)
+        db.mark_annotation_reviewed(current_annotation_id, current_user.username)
         adjudication_handler._adjudicate_annotation()
     elif action == 'comment':
         # No additional changes to be made if only
@@ -728,8 +722,7 @@ def adjudicate_records():
                                            stored_annotation_id)
 
     for annotation_id in annotations_with_duplicates:
-        db.mark_annotation_reviewed(annotation_id, current_user.username,
-                                    ReviewStatus.REVIEWED, ReviewStatus.UNREVIEWED)
+        db.mark_annotation_reviewed(annotation_id, current_user.username)
 
     if len(patient_data["annotation_ids"]) > 0:
             # Only lock the patient for annotation if
